refactor(theme): split appearance parsing and loading modules
