+++ src/ui/theme/appearance/mod.rs
@@ -1,187 +1,30 @@
+mod loading;
+mod parsing;
+mod types;
+
+pub(crate) use self::types::{CodePreviewPalette, Palette};
+use self::{
+    loading::load_theme_from_disk,
+    parsing::normalize_key,
+    types::{
+        ClassStyle, EntryClassCacheKey, PreviewTheme, ResolvedAppearance, RuleOverride, Theme,
+    },
+};
 use super::builtin_themes::DEFAULT_THEME_TOML;
 use crate::{
     app::{Entry, EntryKind, FileClass},
-    config, file_info,
+    file_info,
 };
 use ratatui::style::Color;
-use serde::Deserialize;
 use std::{
     collections::HashMap,
-    fs, io,
-    path::{Path, PathBuf},
+    path::Path,
     sync::{Mutex, OnceLock},
     time::{SystemTime, UNIX_EPOCH},
 };
 
 static ACTIVE_THEME: OnceLock<Theme> = OnceLock::new();
 static ENTRY_CLASS_CACHE: OnceLock<Mutex<HashMap<EntryClassCacheKey, FileClass>>> = OnceLock::new();
-
-#[derive(Clone, Copy)]
-pub(crate) struct Palette {
-    pub bg: Color,
-    pub chrome: Color,
-    pub chrome_alt: Color,
-    pub panel: Color,
-    pub panel_alt: Color,
-    pub surface: Color,
-    pub elevated: Color,
-    pub border: Color,
-    pub text: Color,
-    pub muted: Color,
-    pub accent: Color,
-    pub accent_soft: Color,
-    pub accent_text: Color,
-    pub selected_bg: Color,
-    pub selected_border: Color,
-    pub selection_bar: Color,
-    pub sidebar_active: Color,
-    pub button_bg: Color,
-    pub button_disabled_bg: Color,
-    pub path_bg: Color,
-}
-
-#[derive(Clone, Copy)]
-pub(crate) struct CodePreviewPalette {
-    pub fg: Color,
-    pub bg: Color,
-    pub selection_bg: Color,
-    pub selection_fg: Color,
-    pub caret: Color,
-    pub line_highlight: Color,
-    pub line_number: Color,
-    pub comment: Color,
-    pub string: Color,
-    pub constant: Color,
-    pub keyword: Color,
-    pub function: Color,
-    pub r#type: Color,
-    pub parameter: Color,
-    pub tag: Color,
-    pub operator: Color,
-    pub r#macro: Color,
-    pub invalid: Color,
-}
-
-#[derive(Clone, Copy)]
-struct PreviewTheme {
-    code: CodePreviewPalette,
-}
-
-#[derive(Clone)]
-struct ClassStyle {
-    icon: String,
-    color: Color,
-}
-
-#[derive(Clone, Default)]
-struct RuleOverride {
-    class: Option<FileClass>,
-    icon: Option<String>,
-    color: Option<Color>,
-}
-
-#[derive(Clone)]
-struct Theme {
-    palette: Palette,
-    preview: PreviewTheme,
-    classes: HashMap<FileClass, ClassStyle>,
-    extensions: HashMap<String, RuleOverride>,
-    files: HashMap<String, RuleOverride>,
-    directories: HashMap<String, RuleOverride>,
-}
-
-pub(crate) struct ResolvedAppearance<'a> {
-    #[cfg(test)]
-    pub class: FileClass,
-    pub icon: &'a str,
-    pub color: Color,
-}
-
-#[derive(Clone, Debug, Eq, Hash, PartialEq)]
-struct EntryClassCacheKey {
-    path: PathBuf,
-    is_dir: bool,
-    size: u64,
-    modified: Option<(u64, u32)>,
-}
-
-#[derive(Deserialize, Default)]
-struct ThemeFile {
-    palette: Option<PaletteOverride>,
-    preview: Option<PreviewOverride>,
-    classes: Option<HashMap<String, ClassStyleOverride>>,
-    extensions: Option<HashMap<String, RuleOverrideDef>>,
-    files: Option<HashMap<String, RuleOverrideDef>>,
-    directories: Option<HashMap<String, RuleOverrideDef>>,
-}
-
-#[derive(Deserialize, Default)]
-struct PaletteOverride {
-    bg: Option<String>,
-    chrome: Option<String>,
-    chrome_alt: Option<String>,
-    panel: Option<String>,
-    panel_alt: Option<String>,
-    surface: Option<String>,
-    elevated: Option<String>,
-    border: Option<String>,
-    text: Option<String>,
-    muted: Option<String>,
-    accent: Option<String>,
-    accent_soft: Option<String>,
-    accent_text: Option<String>,
-    selected_bg: Option<String>,
-    selected_border: Option<String>,
-    selection_bar: Option<String>,
-    sidebar_active: Option<String>,
-    button_bg: Option<String>,
-    button_disabled_bg: Option<String>,
-    path_bg: Option<String>,
-}
-
-#[derive(Deserialize, Default)]
-struct PreviewOverride {
-    code: Option<CodePreviewOverride>,
-}
-
-#[derive(Deserialize, Default)]
-struct CodePreviewOverride {
-    fg: Option<String>,
-    bg: Option<String>,
-    selection_bg: Option<String>,
-    selection_fg: Option<String>,
-    caret: Option<String>,
-    line_highlight: Option<String>,
-    line_number: Option<String>,
-    comment: Option<String>,
-    string: Option<String>,
-    constant: Option<String>,
-    keyword: Option<String>,
-    function: Option<String>,
-    r#type: Option<String>,
-    parameter: Option<String>,
-    tag: Option<String>,
-    operator: Option<String>,
-    r#macro: Option<String>,
-    invalid: Option<String>,
-}
-
-#[derive(Deserialize, Default)]
-struct ClassStyleOverride {
-    icon: Option<String>,
-    color: Option<String>,
-}
-
-#[derive(Deserialize)]
-#[serde(untagged)]
-enum RuleOverrideDef {
-    Class(String),
-    Rich {
-        class: Option<String>,
-        icon: Option<String>,
-        color: Option<String>,
-    },
-}
 
 pub(crate) fn initialize() {
     let _ = ACTIVE_THEME.get_or_init(load_theme_from_disk);
@@ -215,38 +58,6 @@ fn active_theme() -> &'static Theme {
 
 fn entry_class_cache() -> &'static Mutex<HashMap<EntryClassCacheKey, FileClass>> {
     ENTRY_CLASS_CACHE.get_or_init(|| Mutex::new(HashMap::new()))
-}
-
-fn load_theme_from_disk() -> Theme {
-    let Some(path) = theme_path() else {
-        return Theme::default_theme();
-    };
-    let contents = match fs::read_to_string(&path) {
-        Ok(contents) => contents,
-        Err(error) if error.kind() == io::ErrorKind::NotFound => return Theme::default_theme(),
-        Err(error) => {
-            eprintln!(
-                "elio: failed to read theme from {}: {error}",
-                path.display()
-            );
-            return Theme::default_theme();
-        }
-    };
-
-    match Theme::from_config_str(&contents) {
-        Ok(theme) => theme,
-        Err(error) => {
-            eprintln!(
-                "elio: failed to load theme from {}: {error}",
-                path.display()
-            );
-            Theme::default_theme()
-        }
-    }
-}
-
-fn theme_path() -> Option<PathBuf> {
-    config::config_dir().map(|dir| dir.join("theme.toml"))
 }
 
 impl Theme {
@@ -1309,54 +1120,6 @@ impl Theme {
         }
     }
 
-    fn from_config_str(config: &str) -> anyhow::Result<Self> {
-        Self::apply_config_on(Self::default_theme(), config)
-    }
-
-    fn apply_config_on(mut theme: Self, config: &str) -> anyhow::Result<Self> {
-        let parsed: ThemeFile = toml::from_str(config)?;
-        theme.apply_overrides(parsed)?;
-        Ok(theme)
-    }
-
-    fn apply_overrides(&mut self, parsed: ThemeFile) -> anyhow::Result<()> {
-        if let Some(palette) = parsed.palette {
-            apply_palette_overrides(&mut self.palette, palette)?;
-        }
-        if let Some(preview) = parsed.preview {
-            apply_preview_overrides(&mut self.preview, preview)?;
-        }
-
-        if let Some(classes) = parsed.classes {
-            for (name, override_style) in classes {
-                let class = parse_class_name(&name)
-                    .ok_or_else(|| anyhow::anyhow!("unknown class `{name}`"))?;
-                let style = self
-                    .classes
-                    .entry(class)
-                    .or_insert_with(|| default_class_style(class));
-                if let Some(icon) = override_style.icon {
-                    style.icon = icon;
-                }
-                if let Some(color) = override_style.color {
-                    style.color = parse_color(&color)?;
-                }
-            }
-        }
-
-        if let Some(extensions) = parsed.extensions {
-            apply_rule_map(&mut self.extensions, extensions)?;
-        }
-        if let Some(files) = parsed.files {
-            apply_rule_map(&mut self.files, files)?;
-        }
-        if let Some(directories) = parsed.directories {
-            apply_rule_map(&mut self.directories, directories)?;
-        }
-
-        Ok(())
-    }
-
     fn resolve(&self, path: &Path, kind: EntryKind) -> ResolvedAppearance<'_> {
         let builtin_class = builtin_classify_path(path, kind);
         self.resolve_with_builtin_class(path, kind, builtin_class)
@@ -1423,112 +1186,6 @@ impl Theme {
             icon,
             color,
         }
-    }
-}
-
-fn apply_palette_overrides(
-    palette: &mut Palette,
-    overrides: PaletteOverride,
-) -> anyhow::Result<()> {
-    apply_palette_color(&mut palette.bg, overrides.bg)?;
-    apply_palette_color(&mut palette.chrome, overrides.chrome)?;
-    apply_palette_color(&mut palette.chrome_alt, overrides.chrome_alt)?;
-    apply_palette_color(&mut palette.panel, overrides.panel)?;
-    apply_palette_color(&mut palette.panel_alt, overrides.panel_alt)?;
-    apply_palette_color(&mut palette.surface, overrides.surface)?;
-    apply_palette_color(&mut palette.elevated, overrides.elevated)?;
-    apply_palette_color(&mut palette.border, overrides.border)?;
-    apply_palette_color(&mut palette.text, overrides.text)?;
-    apply_palette_color(&mut palette.muted, overrides.muted)?;
-    apply_palette_color(&mut palette.accent, overrides.accent)?;
-    apply_palette_color(&mut palette.accent_soft, overrides.accent_soft)?;
-    apply_palette_color(&mut palette.accent_text, overrides.accent_text)?;
-    apply_palette_color(&mut palette.selected_bg, overrides.selected_bg)?;
-    apply_palette_color(&mut palette.selected_border, overrides.selected_border)?;
-    apply_palette_color(&mut palette.selection_bar, overrides.selection_bar)?;
-    apply_palette_color(&mut palette.sidebar_active, overrides.sidebar_active)?;
-    apply_palette_color(&mut palette.button_bg, overrides.button_bg)?;
-    apply_palette_color(
-        &mut palette.button_disabled_bg,
-        overrides.button_disabled_bg,
-    )?;
-    apply_palette_color(&mut palette.path_bg, overrides.path_bg)?;
-    Ok(())
-}
-
-fn apply_palette_color(target: &mut Color, value: Option<String>) -> anyhow::Result<()> {
-    if let Some(value) = value {
-        *target = parse_color(&value)?;
-    }
-    Ok(())
-}
-
-fn apply_preview_overrides(
-    preview: &mut PreviewTheme,
-    overrides: PreviewOverride,
-) -> anyhow::Result<()> {
-    if let Some(code) = overrides.code {
-        apply_code_preview_overrides(&mut preview.code, code)?;
-    }
-    Ok(())
-}
-
-fn apply_code_preview_overrides(
-    code: &mut CodePreviewPalette,
-    overrides: CodePreviewOverride,
-) -> anyhow::Result<()> {
-    apply_palette_color(&mut code.fg, overrides.fg)?;
-    apply_palette_color(&mut code.bg, overrides.bg)?;
-    apply_palette_color(&mut code.selection_bg, overrides.selection_bg)?;
-    apply_palette_color(&mut code.selection_fg, overrides.selection_fg)?;
-    apply_palette_color(&mut code.caret, overrides.caret)?;
-    apply_palette_color(&mut code.line_highlight, overrides.line_highlight)?;
-    apply_palette_color(&mut code.line_number, overrides.line_number)?;
-    apply_palette_color(&mut code.comment, overrides.comment)?;
-    apply_palette_color(&mut code.string, overrides.string)?;
-    apply_palette_color(&mut code.constant, overrides.constant)?;
-    apply_palette_color(&mut code.keyword, overrides.keyword)?;
-    apply_palette_color(&mut code.function, overrides.function)?;
-    apply_palette_color(&mut code.r#type, overrides.r#type)?;
-    apply_palette_color(&mut code.parameter, overrides.parameter)?;
-    apply_palette_color(&mut code.tag, overrides.tag)?;
-    apply_palette_color(&mut code.operator, overrides.operator)?;
-    apply_palette_color(&mut code.r#macro, overrides.r#macro)?;
-    apply_palette_color(&mut code.invalid, overrides.invalid)?;
-    Ok(())
-}
-
-fn apply_rule_map(
-    target: &mut HashMap<String, RuleOverride>,
-    source: HashMap<String, RuleOverrideDef>,
-) -> anyhow::Result<()> {
-    for (key, value) in source {
-        target.insert(normalize_key(&key), parse_rule_override(value)?);
-    }
-    Ok(())
-}
-
-fn parse_rule_override(value: RuleOverrideDef) -> anyhow::Result<RuleOverride> {
-    match value {
-        RuleOverrideDef::Class(class) => {
-            Ok(rule_class(parse_class_name(&class).ok_or_else(|| {
-                anyhow::anyhow!("unknown class `{class}`")
-            })?))
-        }
-        RuleOverrideDef::Rich { class, icon, color } => Ok(RuleOverride {
-            class: match class {
-                Some(class) => Some(
-                    parse_class_name(&class)
-                        .ok_or_else(|| anyhow::anyhow!("unknown class `{class}`"))?,
-                ),
-                None => None,
-            },
-            icon,
-            color: match color {
-                Some(color) => Some(parse_color(&color)?),
-                None => None,
-            },
-        }),
     }
 }
 
@@ -1652,40 +1309,6 @@ fn fingerprint_time(modified: Option<SystemTime>) -> Option<(u64, u32)> {
         .map(|duration| (duration.as_secs(), duration.subsec_nanos()))
 }
 
-fn parse_class_name(name: &str) -> Option<FileClass> {
-    match normalize_key(name).as_str() {
-        "directory" | "dir" | "folder" => Some(FileClass::Directory),
-        "code" => Some(FileClass::Code),
-        "config" => Some(FileClass::Config),
-        "document" | "doc" | "text" => Some(FileClass::Document),
-        "license" | "licence" | "legal" => Some(FileClass::License),
-        "image" | "img" => Some(FileClass::Image),
-        "audio" => Some(FileClass::Audio),
-        "video" => Some(FileClass::Video),
-        "archive" | "compressed" => Some(FileClass::Archive),
-        "font" => Some(FileClass::Font),
-        "data" => Some(FileClass::Data),
-        "file" | "plain" => Some(FileClass::File),
-        _ => None,
-    }
-}
-
-fn normalize_key(value: &str) -> String {
-    value.trim().to_ascii_lowercase()
-}
-
-fn parse_color(value: &str) -> anyhow::Result<Color> {
-    let hex = value.trim().trim_start_matches('#');
-    if hex.len() != 6 {
-        anyhow::bail!("invalid color {value}");
-    }
-
-    let red = u8::from_str_radix(&hex[0..2], 16)?;
-    let green = u8::from_str_radix(&hex[2..4], 16)?;
-    let blue = u8::from_str_radix(&hex[4..6], 16)?;
-    Ok(rgb(red, green, blue))
-}
-
 fn rgb(red: u8, green: u8, blue: u8) -> Color {
     Color::Rgb(red, green, blue)
 }
@@ -1697,6 +1320,7 @@ mod tests {
         env,
         ffi::OsString,
         fs,
+        path::PathBuf,
         sync::{Mutex, OnceLock},
         time::{SystemTime, UNIX_EPOCH},
     };
@@ -1788,15 +1412,15 @@ mod tests {
 
     fn alternate_example_theme_config(name: &str) -> &'static str {
         match name {
-            "default-light" => include_str!("../../../examples/themes/default-light/theme.toml"),
-            "blush-light" => include_str!("../../../examples/themes/blush-light/theme.toml"),
-            "amber-dusk" => include_str!("../../../examples/themes/amber-dusk/theme.toml"),
+            "default-light" => include_str!("../../../../examples/themes/default-light/theme.toml"),
+            "blush-light" => include_str!("../../../../examples/themes/blush-light/theme.toml"),
+            "amber-dusk" => include_str!("../../../../examples/themes/amber-dusk/theme.toml"),
             "catppuccin-mocha" => {
-                include_str!("../../../examples/themes/catppuccin-mocha/theme.toml")
+                include_str!("../../../../examples/themes/catppuccin-mocha/theme.toml")
             }
-            "tokyo-night" => include_str!("../../../examples/themes/tokyo-night/theme.toml"),
-            "navi" => include_str!("../../../examples/themes/navi/theme.toml"),
-            "neon-cherry" => include_str!("../../../examples/themes/neon-cherry/theme.toml"),
+            "tokyo-night" => include_str!("../../../../examples/themes/tokyo-night/theme.toml"),
+            "navi" => include_str!("../../../../examples/themes/navi/theme.toml"),
+            "neon-cherry" => include_str!("../../../../examples/themes/neon-cherry/theme.toml"),
             _ => panic!("unknown alternate example theme fixture: {name}"),
         }
     }
